refactor(docs): separate editing operations and refresh architecture guide (#1127)
* refactor(docs): separate editing operations and refresh architecture guide

* docs: fix architecture links in the published site
